backend/pyproject.toml
[project]
name = "launchpad42-backend"
version = "0.1.0"
description = "42 LaunchPad — Piscine preparation platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "firebase-admin>=6.6.0",
    "google-cloud-firestore>=2.19.0",
    "python-telegram-bot>=21.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "structlog>=24.4.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
